@@ -2,8 +2,8 @@
   "name": "uswagon-auth",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^16.2.12",
-    "@angular/core": "^16.2.12"
+    "@angular/common": "^17.3.10",
+    "@angular/core": "^17.3.10"
   },
   "dependencies": {
     "tslib": "^2.3.0"
