@@ -2,18 +2,14 @@
   "name": "uswagon-core",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^15.2.0",
-    "@angular/core": "^15.2.0"
+    "@angular/common": "^17.3.10",
+    "@angular/core": "^17.3.10"
   },
   "dependencies": {
     "tslib": "^2.3.0"
   },
   "sideEffects": false,
-  "module": "fesm2015/uswagon-core.mjs",
-  "es2020": "fesm2020/uswagon-core.mjs",
-  "esm2020": "esm2020/uswagon-core.mjs",
-  "fesm2020": "fesm2020/uswagon-core.mjs",
-  "fesm2015": "fesm2015/uswagon-core.mjs",
+  "module": "fesm2022/uswagon-core.mjs",
   "typings": "index.d.ts",
   "exports": {
     "./package.json": {
@@ -21,11 +17,9 @@
     },
     ".": {
       "types": "./index.d.ts",
-      "esm2020": "./esm2020/uswagon-core.mjs",
-      "es2020": "./fesm2020/uswagon-core.mjs",
-      "es2015": "./fesm2015/uswagon-core.mjs",
-      "node": "./fesm2015/uswagon-core.mjs",
-      "default": "./fesm2020/uswagon-core.mjs"
+      "esm2022": "./esm2022/uswagon-core.mjs",
+      "esm": "./esm2022/uswagon-core.mjs",
+      "default": "./fesm2022/uswagon-core.mjs"
     }
   }
 }
